Double quotes are not properly escaped
Bazel passes to CFLAGS and CXXFLAGS the redacted values of __DATE__,
__TIMESTAMP__, and __TIME__.

Just a single backslash is not enough. It gets passed into the code
as redacted (without any quotes) and not "redacted". The compiler
complains about the symbol "redacted" not being found for packages
that actually use on of those fields.

Remove hack in apache_httpd/BUILD that was put in place because of this
bug.

Also need to turn off escaping on heredoc when generating
crosstool_bazel.cmake. Without the single quotes around EOF, the
contents of the heredoc are escaped. With single quotes, they are not.

https://stackoverflow.com/a/9870274