Replace 'target' attribute with 'targets'.
This is so that multiple labels could be provided to a toolchain as
sources. An example is `emcmake`, which needs both `emscripten` and
`binaryen` to run.

Another reason for this is to pass in config files that might be needed
for the toolchain. The config files can be passed in as a label and will
be available in ${EXT_BUILD_DEPS}/bin, just like any other directory.