Improve failed exports logic
If something is misconfigured in your setup, these commands can fail.
Example:

```
xcodebuild: error: SDK "iphonesimulator14.4" cannot be located.
xcodebuild: error: SDK "iphonesimulator14.4" cannot be located.
xcrun: error: unable to lookup item 'Path' in SDK 'iphonesimulator14.4
```

In this case, because the execution of the command, and the `export`,
were in the statement, bash ignores the exit code. This breaks that up.

We could also break this up with:

```
foo=$(...)
export foo
```

But that doesn't work in the postprocessing currently done with exports.