Expand make variables in env
This allows toolchains to bring along files and for those files to be
referenced during the build.

This is a slightly breaking change - any $s currently present in env var
attributes which are not consumed by expand_location will need to be
escaped with a second $, because Make-variable expansion requires that
escaping. See the example change done to examples/third_party/openssl/BUILD.openssl.bazel