Use expand_locations_and_make_variables everywhere
expand_locations leads to broken commandlines if used by itself since it
expects dollar signs to be escaped, but does not unescape them.

This is a mildly breaking change as previously accepted uses of $ now
have to be replaced by $$. As a benefit, this escaping scheme is now
consistent across all attributes as well as with Bazel native rules.