Faciliate usage of jom
The jom tool (a fork of nmake which supports parallel builds) does not
support the "-C <dir>" arg. Given that the build script generated by
rules_foreign_cc changes the working directory to $BUILD_TMPDIR before
invoking make/nmake/jom, there is no need to add the "-C $BUILD_TMPDIR"
arg. This commit removes the addition of the "-C $BUILD_TMPDIR" arg so
that jom can be used as a make variant.