Fix current_*_toolchains
The "Make Variables" defined for the current_*_toolchains contained
"$EXT_BUILD_ROOT/" in the path to the tool. As the current_*_toolchains
are intended to be used in non foreign_cc rules, EXT_BUILD_ROOT is
undefined and the tool path is incorrect.

This commit resolves this issue and adds a test to CI