Allow selecting individual outputs from a rule
If one wants to select a single output from a build (like a single
shared library or a single static library), it's not possible to do so
with the current code. The current code only creates an output_group for
the declared binaries, which isn't sufficient.

This change creates an output group for all the artifacts including the
includes.

Fixes #768