fix: let 'toolchain' attr be selectable
The toolchain attr is not currently selectable making builds that need
to handle multiple platforms (incl windows) very verbose. The target
needs to be declared at least twice, once for windows and once for unix
based platforms.

By making it selectable, even though we still need a rule() and
rule_variant() declaration, we can wrap it in a macro and only expose
the macro to the user.