cmake: put CMAKE_SYSTEM_{NAME,PROCESSOR} into toolchain file
For some projects, it's sufficient to just have these in the cache, but
for abseil, it appears to break unless they're in the toolchain file.
The error is similar to the one in [this
bug](https://github.com/abseil/abseil-cpp/issues/1241) where it tries to
use x64 flags for an aarch64 target.

The [cmake
docs](https://cmake.org/cmake/help/book/mastering-cmake/chapter/Cross%20Compiling%20With%20CMake.html#toolchain-files)
also indicate they're supposed to be in the toolchain file, so this
aligns there, too.