feat: support make vars in postfix_script
postfix_script did not support make vars which means we cannot use
things like $(execpath :target) in it. This limited the functionality of
the attribute.

Note, this is a breaking change as $$ wrapped variable declaration and
single $ shell variable declaration will no longer work in
postfix_script e.g. $$INSTALLDIR$$ and $INSTALLDIR. Need to used shell
variable with an escaped $ like $$INSTALLDIR.