fix: broken msvc builds
Using "\"$EXT_BUILD_ROOT\"/" with the extra \"s is unnecessary. The extra
quote should be applied to string where this is used. Otherwise this
causes odd path problems for msvc.

Notably, in get_env_prelude, this causes LIB and INCLUDE env vars to
expand to something like:
LIB="\\\\\"$EXT_BUILD_ROOT\\\\\"/external/toolchain..."
This is not recognized as a valid path when running through msys2 on
windows and causes all kinds of pain.