Fix target name for meson_with_requirements
meson_with_requirements has a small quirk where the resulting rule name ends up with an '_' postfixed to the end. This is unexpected and a bit annoying when the meson project is a dependency of another rule since you have to remember to add the _ in any include paths etc.