cmake: emit warning when failing to cross-compile
cmake requires setting CMAKE_SYSTEM_NAME to detect a cross-compile of
any form (including same-os but different-processor). This means that,
if an OS is not listed under _TARGET_OS_PARAMS, cross-compilation can
never work on that platform; it would have silently done a native
compile.

This change makes it emit a warning in this case, and adds macos to the
list so xcompiles work on macos.